import data via pravega client
