+++ scenarios/pravega-flink-connector-sql-samples/src/main/python/common.py
@@ -7,7 +7,6 @@ You may obtain a copy of the License at
 
   http://www.apache.org/licenses/LICENSE-2.0
 """
-
 import argparse
 
 DEFAULT_SCOPE = 'taxi'
